% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_plot.R
\name{barcode_plot}
\alias{barcode_plot}
\title{Plot Protein Barcodes For Proteins Of Interest}
\usage{
barcode_plot(
  data,
  protein_id,
  proteins,
  comparison,
  comparisons = NULL,
  diff,
  start,
  end,
  length,
  colour
)
}
\arguments{
\item{data}{Data frame containing peptide differential abundance results}

\item{protein_id}{Column containing protein IDs.}

\item{proteins}{Protein(s) to plot. No limit, but more than 5 or so will look messy.}

\item{comparison}{Column containing the comparison.}

\item{comparisons}{Comparison(s) to plot. All by default.}

\item{diff}{Column containing differential abundance data.}

\item{start}{Column containing peptide start positions.}

\item{end}{Column containing peptide end positions.}

\item{length}{Column containing protein lengths.}

\item{colour}{Column containing colour indications. Should have three levels: "down", "unchanged" and "up".}
}
\value{
Barcode plots for the selected proteins and comparisons, indicating peptides that have changed in abundance between samples.
}
\description{
Plot Protein Barcodes For Proteins Of Interest
}
