% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_correction.R
\name{bh_correction}
\alias{bh_correction}
\title{Perform Multiple Testing Correction Using The BH Procedure}
\usage{
bh_correction(input, correction, alpha)
}
\arguments{
\item{input}{A data frame containing the output from the Wasserstein test
for a given comparison.}

\item{correction}{Logical indicating whether or not to correct the
p-values (using the BH procedure).}

\item{alpha}{The significance level to use.}
}
\value{
A corrected data frame containing the output from the Wasserstein
test.
}
\description{
Perform Multiple Testing Correction Using The BH Procedure
}
