% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_dist_1.R
\name{int_dist_1}
\alias{int_dist_1}
\title{Add Metadata To A Peptide Intensity Distribution}
\usage{
int_dist_1(input)
}
\arguments{
\item{input}{A data frame generated by peptide_intensity_distribution,
containing peptide-level data for one protein and one experiment.}
}
\value{
A list containing cumulative peptide intensity distributions
and associated metadata.
}
\description{
Add Metadata To A Peptide Intensity Distribution
}
