% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_dist_2.R
\name{int_dist_2}
\alias{int_dist_2}
\title{Generate A Cumulative Peptide Intensity Distribution}
\usage{
int_dist_2(input)
}
\arguments{
\item{input}{A data frame generated by peptide_intensity_distribution,
containing peptide-level data for one protein and one experiment.}
}
\value{
A data frame containing a cumulative peptide intensity
distribution.
}
\description{
Generate A Cumulative Peptide Intensity Distribution
}
