% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_distribution.R
\name{intensity_distribution}
\alias{intensity_distribution}
\title{Calculate An Individual Peptide Intensity Distribution}
\usage{
intensity_distribution(data, value, start, end, peptide_length, protein_length)
}
\arguments{
\item{data}{A data frame containing peptide-level data for a single protein and experiment. Comes from a nested tibble for the peptide_intensity_distribution function. All other arguments are inherited from peptide_intensity_distribution.}

\item{value}{Column containing peptide intensities.}

\item{start}{Column containing start residue of peptide.}

\item{end}{Column containing end residue of peptide.}

\item{peptide_length}{Column containing peptide length.}

\item{protein_length}{Column containing protein length.}
}
\value{
A tibble with four columns containing the peptide intensity distribution for the protein and experiment.
}
\description{
Calculate An Individual Peptide Intensity Distribution
}
