% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_int_dist_1.R
\name{mean_int_dist_1}
\alias{mean_int_dist_1}
\title{Add Metadata To A Mean Peptide Intensity Distribution}
\usage{
mean_int_dist_1(input)
}
\arguments{
\item{input}{A list containing cumulative peptide intensity distributions
for multiple replicates for one protein and condition.}
}
\value{
A list containing a mean cumulative peptide intensity distribution
and its associated metadata.
}
\description{
Add Metadata To A Mean Peptide Intensity Distribution
}
