% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_int_dist_2.R
\name{mean_int_dist_2}
\alias{mean_int_dist_2}
\title{Generate A Mean Cumulative Peptide Intensity Distribution}
\usage{
mean_int_dist_2(input)
}
\arguments{
\item{input}{A list containing cumulative peptide intensity distributions
for multiple replicates for one protein and condition.}
}
\value{
A data frame containing a mean cumulative peptide intensity
distribution.
}
\description{
Generate A Mean Cumulative Peptide Intensity Distribution
}
