% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_peptide_intensity_distribution.R
\name{mean_peptide_intensity_distribution}
\alias{mean_peptide_intensity_distribution}
\title{Calculate Mean Peptide Intensity Distributions}
\usage{
mean_peptide_intensity_distribution(data)
}
\arguments{
\item{data}{A multi-level list (protein, condition, replicate) containing
cumulative peptide intensity distributions and their associated metadata.
For example, the output from peptide_intensity_distribution.}
}
\value{
A multi-level list in the same format as the input, containing mean
cumulative peptide intensity distributions for each protein and
condition along with metadata.
}
\description{
Calculate Mean Peptide Intensity Distributions
}
