% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{moment}
\alias{moment}
\title{Calculate Peptide Intensity Moments}
\usage{
moment(
  data,
  protein_id,
  experiment,
  value,
  position,
  length,
  sep = "/",
  anchor = "centre",
  normalise = TRUE
)
}
\arguments{
\item{data}{A data frame containing peptide-level proteomics data.}

\item{protein_id}{Column containing protein names.}

\item{experiment}{Column containing experiment.}

\item{value}{Column containing peptide intensity data.}

\item{position}{Column containing peptide midpoint.}

\item{length}{Column containing protein length.}

\item{sep}{Separating character. Must not be present in either protein names
or experiment. Default = "/".}

\item{anchor}{Must be set to "centre" (default), "N" or "C". Defines which
position in the protein the moment is relative to.}

\item{normalise}{Logical indicating whether the position should be
normalised to the protein length (default = TRUE).}
}
\value{
A data frame containing peptide intensity moments for each protein
and sample.
}
\description{
Calculate Peptide Intensity Moments
}
