% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_simulation.R
\name{moment_simulation}
\alias{moment_simulation}
\title{Simulate protein moments by sampling peptides}
\usage{
moment_simulation(data, number_of_peptides, iterations = 100)
}
\arguments{
\item{data}{Peptide data to use for simulations}

\item{number_of_peptides}{Number of peptides to use for simulations}

\item{iterations}{Number of iterations to perform}
}
\value{
A vector of simulated moments (mean value across all of the )
}
\description{
Simulate protein moments by sampling peptides
}
