% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_test.R
\name{moment_test}
\alias{moment_test}
\title{Statistical testing of protein moments}
\usage{
moment_test(
  moment_data,
  peptide_data,
  proteins = NULL,
  test_sample,
  reference_sample,
  replicate_threshold = 2,
  peptide_threshold = 1,
  iterations = 100
)
}
\arguments{
\item{moment_data}{Data frame containing moments (for individual replicates)}

\item{peptide_data}{Data frame containing all peptides}

\item{proteins}{Vector of protein(s) to analyse (default is all proteins in data)}

\item{test_sample}{Sample to test}

\item{reference_sample}{Sample to use as reference}

\item{replicate_threshold}{Number of replicates required}

\item{peptide_threshold}{Number of peptides required}

\item{iterations}{Number of iterations of simulation to perform}
}
\value{
Data frame containing test results
}
\description{
Statistical testing of protein moments
}
