% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_test_single.R
\name{moment_test_single}
\alias{moment_test_single}
\title{Statistical testing of moment for single protein}
\usage{
moment_test_single(
  moment_data,
  peptide_data,
  protein,
  test_sample,
  reference_sample,
  replicate_threshold = 2,
  peptide_threshold = 1,
  iterations = 100
)
}
\arguments{
\item{moment_data}{Data frame containing moments (for individual replicates)}

\item{peptide_data}{Data frame containing all peptides}

\item{protein}{Single protein to test}

\item{test_sample}{Sample to test}

\item{reference_sample}{Sample to use as reference}

\item{replicate_threshold}{Number of replicates required}

\item{peptide_threshold}{Number of peptides required}

\item{iterations}{Number of iterations of simulation to perform}
}
\value{
Data frame with one row, containing test result
}
\description{
Statistical testing of moment for single protein
}
