% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_intensity_distribution.R
\name{peptide_intensity_distribution}
\alias{peptide_intensity_distribution}
\title{Generate Cumulative Peptide Intensity Distributions}
\usage{
peptide_intensity_distribution(
  data,
  names_col,
  exp_col,
  cond_col,
  rep_col,
  seq_col,
  val_col,
  start_col,
  end_col,
  len_col,
  sep = "/"
)
}
\arguments{
\item{data}{A data frame containing peptide-level proteomics data.}

\item{names_col}{Column containing protein names.}

\item{exp_col}{Column containing experiment.}

\item{cond_col}{Column containing condition.}

\item{rep_col}{Column containing replicate.}

\item{seq_col}{Column containing peptide sequence.}

\item{val_col}{Column containing peptide intensity data.}

\item{start_col}{Column containing start residue of peptide.}

\item{end_col}{Column containing end residue of peptide.}

\item{len_col}{Column containing protein length.}

\item{sep}{Separating character. Must not be present in either protein names
or experiment.}
}
\value{
A multi-level list containing cumulative peptide intensity
distributions (cumulative proportion of intensity observed against
cumulative proportion of N-to-C distance). The list is separated
first by protein, then by condition, then by replicate.
}
\description{
Generate Cumulative Peptide Intensity Distributions
}
