% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensity_distribution.R
\name{plot_intensity_distribution}
\alias{plot_intensity_distribution}
\title{Plot Cumulative Peptide Intensity Distributions}
\usage{
plot_intensity_distribution(
  data,
  protein,
  condition,
  midpoint = FALSE,
  ribbon = FALSE
)
}
\arguments{
\item{data}{A multi-level list containing cumulative peptide intensity
distributions, either for individual replicates or means.}

\item{protein}{A single protein name to plot.}

\item{condition}{A vector of conditions to plot.}

\item{midpoint}{Logical indicating whether or not to plot the midpoints of
the cumulative peptide intensity distributions.}

\item{ribbon}{Logical indicating whether to plot a ribbon showing mean
plus/minus standard deviation. Only applicable if plotting mean
distributions.}
}
\value{
A plot of cumulative peptide intensity against position in protein
(from N terminus to C terminus) for the protein of interest. If
midpoint = TRUE, midpoints are printed to the console.
}
\description{
Plot Cumulative Peptide Intensity Distributions
}
