% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_midpoint.R
\name{plot_midpoint}
\alias{plot_midpoint}
\title{Plot The Midpoint Of A Cumulative Peptide Intensity Distribution}
\usage{
plot_midpoint(input, plot, shapes)
}
\arguments{
\item{input}{A data frame from plot_intensity_distribution, containing
cumulative peptide intensity distributions for the protein, condition(s)
and replicate(s) of interest.}

\item{plot}{A ggplot2 object created by plot_intensity_distribution.}

\item{shapes}{A vector of point shapes, defined in
plot_intensity_distribution.}
}
\value{
A ggplot2 object containing points plotted where 50\% of the
cumulative intensity is reached for each distribution.
}
\description{
Plot The Midpoint Of A Cumulative Peptide Intensity Distribution
}
