% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wass_stat_extended.R
\name{wass_stat_extended}
\alias{wass_stat_extended}
\title{Perform Wasserstein Test On Actual Peptide Intensity Distributions}
\usage{
wass_stat_extended(vec1, vec2)
}
\arguments{
\item{vec1}{A peptide intensity distribution for the first experiment (reference).}

\item{vec2}{A peptide intensity distribution for the second experiment (sample).}
}
\value{
A vector containing multiple difference metrics from the comparison performed.
}
\description{
Perform Wasserstein Test On Actual Peptide Intensity Distributions
}
