% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wass_stat_permutation.R
\name{wass_stat_permutation}
\alias{wass_stat_permutation}
\title{Perform Wasserstein Test On Permutations Of A Peptide Intensity Distribution}
\usage{
wass_stat_permutation(vec1, vec2)
}
\arguments{
\item{vec1}{A peptide intensity distribution for the first permutation ("reference").}

\item{vec2}{A peptide intensity distribution for the second permutation ("sample").}
}
\value{
The absolute difference between the two permutations.
}
\description{
Perform Wasserstein Test On Permutations Of A Peptide Intensity Distribution
}
