% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_test.R
\name{wasserstein_test}
\alias{wasserstein_test}
\title{Statistical Testing Of Cumulative Peptide Intensity Distributions}
\usage{
wasserstein_test(
  data,
  ref_condition,
  ref_replicate = "mean",
  nboots = 5000,
  correction = TRUE,
  alpha = 0.01,
  sep = "/"
)
}
\arguments{
\item{data}{A multi-level list (protein, condition, replicate) containing
cumulative peptide intensity distributions and their associated metadata.
These may be individual replicates or means.}

\item{ref_condition}{The condition to use as reference.}

\item{ref_replicate}{The replicate to use as reference (default = "mean").}

\item{nboots}{The number of bootstrap iterations to perform.}

\item{correction}{Logical indicating whether or not to correct the
p-values (using the BH procedure).}

\item{alpha}{The significance level to use.}

\item{sep}{Separating character. Must not be present in experiment names.}
}
\value{
A list of data frames containing the Wasserstein test results for
each comparison.
}
\description{
Statistical Testing Of Cumulative Peptide Intensity Distributions
}
