% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_test_2.R
\name{wasserstein_test_2}
\alias{wasserstein_test_2}
\title{Statistical Testing Of Peptide Intensity Distributions}
\usage{
wasserstein_test_2(data, ref_condition, nboots = 5000, method = "BH")
}
\arguments{
\item{data}{A multi-level list (protein, condition, replicate) containing
cumulative peptide intensity distributions and their associated metadata.
These may be either individual replicates (in which case the
variation between replicates will be taken into account) but may also be
means.}

\item{ref_condition}{The condition to use as reference.}

\item{nboots}{The number of bootstrap iterations to perform.}

\item{method}{P-value correction method to use in p.adjust. Should be one of
p.adjust.methods (default "BH").}
}
\value{
A list of data frames containing the Wasserstein test results for
each comparison.
}
\description{
Statistical Testing Of Peptide Intensity Distributions
}
