% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{moment}
\alias{moment}
\title{Calculate Peptide Intensity Moments}
\usage{
moment(data, names_col, exp_col, val_col, pos_col, len_col, sep = "/")
}
\arguments{
\item{data}{A data frame containing peptide-level proteomics data.}

\item{names_col}{Column containing protein names.}

\item{exp_col}{Column containing experiment.}

\item{val_col}{Column containing peptide intensity data.}

\item{pos_col}{Column containing peptide midpoint.}

\item{len_col}{Column containing protein length.}

\item{sep}{Separating character. Must not be present in either protein names
or experiment.}
}
\value{
A data frame containing peptide intensity moments for each protein
and sample.
}
\description{
Calculate Peptide Intensity Moments
}
