% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wass_test_2.R
\name{wass_test_2}
\alias{wass_test_2}
\title{Compare Peptide Intensity Distributions Using The Wasserstein Test}
\usage{
wass_test_2(
  reference,
  sample,
  nboots = 5000,
  protein,
  ref_condition,
  sam_condition
)
}
\arguments{
\item{reference}{A list containing data for the reference condition.}

\item{sample}{A list containing data for the sample condition.}

\item{nboots}{The number of bootstrap iterations to perform.}

\item{protein}{The protein being tested.}

\item{ref_condition}{The reference condition.}

\item{sam_condition}{The sample condition being tested.}
}
\value{
A data frame containing the output of the Wasserstein test for the
comparison of interest.
}
\description{
Compare Peptide Intensity Distributions Using The Wasserstein Test
}
