% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wass_test_dist.R
\name{wass_test_dist}
\alias{wass_test_dist}
\title{Compare Peptide Intensity Distributions Using The Wasserstein Test}
\usage{
wass_test_dist(reference, sample, nboots)
}
\arguments{
\item{reference}{A data frame containing the reference cumulative peptide
intensity distribution.}

\item{sample}{A data frame containing the sample cumulative peptide
intensity distribution to test.}

\item{nboots}{The number of bootstrap iterations to perform.}
}
\value{
A data frame containing the output of the Wasserstein test for the
comparison of interest.
}
\description{
Compare Peptide Intensity Distributions Using The Wasserstein Test
}
