% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woods_plot.R
\name{woods_plot}
\alias{woods_plot}
\title{Plot Woods Plots for Proteins Of Interest}
\usage{
woods_plot(
  data,
  protein_id,
  proteins,
  comparison,
  comparisons = NULL,
  diff,
  start,
  end,
  length,
  lower,
  upper,
  colour
)
}
\arguments{
\item{data}{Data frame containing peptide differential abundance results}

\item{protein_id}{Column containing protein IDs.}

\item{proteins}{Protein(s) to plot. No limit, but more than 5 or so will look messy.}

\item{comparison}{Column containing the comparison.}

\item{comparisons}{Comparison(s) to plot. All by default.}

\item{diff}{Column containing differential abundance data.}

\item{start}{Column containing peptide start positions.}

\item{end}{Column containing peptide end positions.}

\item{length}{Column containing protein lengths.}

\item{lower}{Column indicating lower limit of 95\% CI.}

\item{upper}{Column indicating lower limit of 95\% CI.}

\item{colour}{Column containing colour indications. Should have three levels: "down", "unchanged" and "up".}
}
\value{
Woods plots for the selected proteins and comparisons, indicating peptides that have changed in abundance between samples.
}
\description{
Plot Woods Plots for Proteins Of Interest
}
